FCN-334 split advisor recommendations and total clusters into focused cards
splits the monolithic AdvisorRecommendations into AdvisorSeverity + AdvisorCategories,
and extracts ClusterProviders from TotalClusters per Figma mockup layout.
