+++ src/components/dashboard/AdvisorRecommendations/AdvisorRecommendations.module.scss
@@ -69,4 +69,3 @@
   font-size: var(--pf-t--global--font--size--sm);
   color: var(--pf-t--global--text--color--regular);
 }
-
