+++ src/components/dashboard/TotalClusters/ClusterProviders.module.scss
@@ -1,0 +1,34 @@
+.container {
+  padding: var(--pf-t--global--spacer--md);
+  gap: var(--pf-t--global--spacer--md);
+}
+
+.chartContainer {
+  width: 160px;
+  height: 160px;
+}
+
+.legendGrid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: var(--pf-t--global--spacer--sm) var(--pf-t--global--spacer--lg);
+}
+
+.legendItem {
+  display: flex;
+  align-items: center;
+  gap: var(--pf-t--global--spacer--xs);
+}
+
+.legendDot {
+  display: inline-block;
+  width: 10px;
+  height: 10px;
+  border-radius: 2px;
+  flex-shrink: 0;
+}
+
+.legendText {
+  font-size: var(--pf-t--global--font--size--sm);
+  color: var(--pf-t--global--text--color--regular);
+}
