+++ src/components/dashboard/TotalClusters/TotalClusters.module.scss
@@ -19,27 +19,6 @@
   color: var(--pf-t--global--text--color--subtle);
 }
 
-.breakdown {
-  padding: var(--pf-t--global--spacer--md) 0;
-}
-
-.breakdownCount {
-  font-size: var(--pf-t--global--font--size--2xl);
-  font-weight: 400;
-  line-height: 1;
-  color: var(--pf-t--global--text--color--regular);
-}
-
-.breakdownLabel {
-  font-size: var(--pf-t--global--font--size--sm);
-  color: var(--pf-t--global--text--color--regular);
-}
-
-.icon {
-  color: var(--pf-t--global--text--color--regular);
-  font-size: var(--pf-t--global--font--size--md);
-}
-
 .viewLink {
   padding-top: var(--pf-t--global--spacer--md);
 }
