+++ src/components/dashboard/TotalClusters/TotalClusters.stories.tsx
@@ -16,81 +16,26 @@ type Story = StoryObj<typeof TotalClusters>;
 
 export const Default: Story = {
   args: {
-    data: {
-      total: 67,
-      breakdown: [
-        { label: 'ROSA', count: 42 },
-        { label: 'ARO', count: 15 },
-        { label: 'OSD', count: 10 },
-      ],
-    },
+    data: { total: 67 },
   },
 };
 
 export const WithViewMore: Story = {
   args: {
-    data: {
-      total: 67,
-      breakdown: [
-        { label: 'ROSA', count: 42 },
-        { label: 'ARO', count: 15 },
-        { label: 'OSD', count: 10 },
-      ],
-    },
+    data: { total: 67 },
     onViewMore: fn(),
-  },
-};
-
-export const TotalOnly: Story = {
-  args: {
-    data: {
-      total: 120,
-    },
-  },
-};
-
-export const SingleClusterType: Story = {
-  args: {
-    data: {
-      total: 25,
-      breakdown: [{ label: 'ROSA', count: 25 }],
-    },
   },
 };
 
 export const ZeroClusters: Story = {
   args: {
-    data: {
-      total: 0,
-      breakdown: [
-        { label: 'ROSA', count: 0 },
-        { label: 'ARO', count: 0 },
-        { label: 'OSD', count: 0 },
-      ],
-    },
+    data: { total: 0 },
   },
 };
 
 export const HighCount: Story = {
   args: {
-    data: {
-      total: 1247,
-      breakdown: [
-        { label: 'ROSA', count: 800 },
-        { label: 'ARO', count: 300 },
-        { label: 'OSD', count: 100 },
-        { label: 'Other', count: 47 },
-      ],
-    },
+    data: { total: 1247 },
     onViewMore: fn(),
-  },
-};
-
-export const EmptyBreakdown: Story = {
-  args: {
-    data: {
-      total: 50,
-      breakdown: [],
-    },
   },
 };
