+++ src/components/dashboard/TotalClusters/index.ts
@@ -1,2 +1,5 @@
 export { TotalClusters } from './TotalClusters';
-export type { TotalClustersProps, TotalClustersData, ClusterBreakdown } from './TotalClusters';
+export type { TotalClustersProps, TotalClustersData } from './TotalClusters';
+
+export { ClusterProviders } from './ClusterProviders';
+export type { ClusterProvidersProps, ProviderBreakdown } from './ClusterProviders';
